完善战斗接口，新增获取某个队伍当前时间战斗中的记录接口，优化获取副本信息接口，优化英雄上下阵操作接口，新增校验英雄是否已在其他队伍上阵方法，新增校验当前队伍是否存英雄重复上阵方法，新增获取英雄战力并校验该英雄是否属于当前玩家方法，新增更新玩家英雄上下阵状态方法，新增战力提升接口，新增英雄装备穿戴or激活接口，优化获取玩家背包信息接口，优化获取英雄详细信息接口

+++ gm-common/src/main/java/com/gm/common/utils/LotteryGiftsUtils.java
@@ -51,7 +51,7 @@ public class LotteryGiftsUtils {
         return sortOrignalRates.indexOf(nextDouble);
     }
 
-    public static Map<Integer, Integer> gifPron(List<Double> orignalRates,long Num){
+    public static Map<Integer, Integer> gifPron(List<Double> orignalRates,int Num){
         // statistics
         Map<Integer, Integer> count = new HashMap<Integer, Integer>();
         //抽奖次数
